SAK-40020 Students with the same name do not appear in list of exceptions
- Extracting common method to order maps by value
- Adding eid to avoid confusion, like requested
